PreferredHeight is Height of Content.

+++ Character_Reporting/src/net/sf/anathema/character/reporting/pdf/rendering/general/box/BoundsEncoder.java
@@ -5,9 +5,9 @@ import net.sf.anathema.character.reporting.pdf.rendering.graphics.SheetGraphics;
 
 public interface BoundsEncoder {
 
-  public static final int HEADER_HEIGHT = 12;
-  public static final int ARC_SPACE = HEADER_HEIGHT / 2;
-  public static final int ARC_SIZE = 2 * ARC_SPACE;
+  public static final float HEADER_HEIGHT = 12;
+  public static final float ARC_SPACE = HEADER_HEIGHT / 2;
+  public static final float ARC_SIZE = 2 * ARC_SPACE;
 
   public void encodeBoxBounds(SheetGraphics graphics, Bounds bounds);
 }
